// libs/api/domains/syslumenn/src/lib/client/models/operatingLicense.ts
export interface IOperatingLicense {
  rowNum: number
  utgefidAf: string
  leyfisnumer: string
  stadur: string
  kallast: string
  gata: string
  postnumer: string
  tegund: string
  gildirTil: string
  leyfishafi: string
  flokkur: string
  leyfi_Til_Utiveitinga: string
  afgr_Afgengis_Virkirdagar: string
  afgr_Afgengis_Adfaranott_Fridaga: string
  afgr_Afgengis_Virkirdagar_Utiveitingar: string
  afgr_Afgengis_Adfaranott_Fridaga_Utiveitingar: string
}